Fixed client destructor
The model for the client was modified to add another action to the
destructor event. The action is used to signal back to the caller
that the destructor has finished. It was broke, until I realized
that I specified the destructor API method to require a reply of
SUCCESS or FAILURE.